SWS: Fix unsupported objects handling
In case of unsupported type object we will get a None
value from the object parser, and we will check if None
and this object already exists in the unsupported object
list, and incase it's  not exists will add it to the list
and continue to the next object.
So what happens first time an unsupported object arrives
we will add it to the unsupported object list and continue
to the next object, but a second unsupported object of the
same type will get None but it will fail the condition
because it is already in the unsupported object list so we
will continue with this None object.

Fixes: d0b37333e0a7 ("Print one error per unsupported object")
Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>