Add ability to use relative file paths
Currently, mlx_steering_dump and mlx_hw_steering_dump only accept
absolute file paths. This commit adds a check to see if the file
specified with -f exists in the current directory and uses that
as the -f input to the corresponding python script if so. If not,
the original content of the -f argument is passed.