HWS: Group related FW STEs under the same rule
FW GTA STEs are a part of the FW GTA emulator, and they consist of:

1. An EMPTY MATCH STE with an empty tag
2. A RANGE MATCH STE

The index of the RANGE MATCH STE (hit index of the EMPTY MATCH STE), and
the A RANGE MATCH STE are within the same matcher address boundaries.
We use that to group them under the same rule.

Signed-off-by: Itamar Gozlan <igozlan@nvidia.com>