HWS: Support dump over DOCA with pid
DPDK is running a server with a socket ready to connect.
When triggered by the application, it will use the DPDK dump tool
only with pid.
This mechanism is implemented for DOCA, and this commit adds the needed
support from the dump tool. When the DOCA context is present,
there is no need to dump the DPDK context.
The README file is also changed, and the code was refactored to
indicate support for both applications.

Signed-off-by: Itamar Gozlan <igozlan@nvidia.com>