HWS: Set the minimum Python version to 3.9
The script now checks the Python version and exits if it's lower than
the minimum supported version. The alternative is cryptic stack traces
that users may not understand - it is inevitable that we use *some*
features that are not available in older versions of Python. The printed
error looks like this:
```
> python38 hws/mlx_hw_steering_parser.py
Error: mlx_hw_steering_parser.py requires Python 3.9 or newer
```

Before this changeset, the minimum supported version was at least 3.6
due to using f-strings. We can validate this by running
`pipx run vermin -vvv hws/mlx_hw_steering_parser.py`.

With this change, we set the minimum to 3.9. This should be
reasonable as 3.9 is currently the oldest supported version. It was
released on 2020-10-05 and will be EOL on 2025-10. The DOCA dev
container is based on Ubuntu 22.04, which has Python 3.10.

Signed-off-by: Alexander Efimov <aefimov@nvidia.com>