HWS: Add function to print object with indentation
Up to this point, object formatting was always been handled by the
tree_print method of each object. This is not ideal because it returns
only a string, which would mean that emitting the object as JSON would
require a 2nd function to duplicate much of the logic in that function.
To prevent that, we want to be able to get the object as a dict, and be
able to format that dict as either JSON (introduced later in this commit
series) or as a string (existing functionality).

This commit adds the first step in that direction: a function that
pretty-prints an object as a string. This function corresponds almost
exactly to the current format of the tree_print method, so it is
suitable as a drop-in replacement. The output is not char-for-char
identical, but it should be close enough that human readers should not
notice a difference.

Signed-off-by: Alexander Efimov <aefimov@nvidia.com>