HWS: Define base class for Printable objects
Add a small base class to keep track of which classes we've converted to
support emitting as JSON.

The dump_obj method provides a single known interface for emitting an
object as a dict. The sturcture of the fields is not the same as the
struture we wish to print (to preserve the existing structure of the
text output), so it has a parameter to control whether to transform the
object to make it suitable for printing. This transformation is done
within the method rather than by the caller since the caller doesn't
have visiblity into the type of the object. Rather than passing around a
function with the returned dict, just do the transformation in the
method itself.

Signed-off-by: Alexander Efimov <aefimov@nvidia.com>