HWS: Add CLI flag to emit JSON
This commit simply emits the existing text-based tree as a string in a
JSON object. The following commits will incrementally convert the
objects inside the parse context from text-only to dictionary-based
representations. The dictionary-based representations can then be used
to either print the object as tree_print does now, or emit the object as
JSON for machine processing.

The text-based output remains identical.

The current JSON output is simply:
```
{"context": "..."}
```

Signed-off-by: Alexander Efimov <aefimov@nvidia.com>