HWS: Make dr_context printable
This will enable formatting the data as JSON. Since dr_context is the
top-level object, update the main() code to call the new dump_obj
function instead of tree_print. It is a drop-in replacement for
tree_print when transform_for_print is true, so this should have almost
no user-visible changes.

In this commit, most objects are emitted as string values in the JSON
dictionary. Only the top-level context is actually emitted as an object.

The current JSON output looks like:
```
{
    "data": {
        "mlx5dr_debug_res_type": "4000",
        "id": "0x55dc5151a4e0",
        "hws_support": "True",
        "dev_name": "mlx5_0",
        "debug_version": "1.0.NV_HWS_25.04"
    },
    "attr": "Attr: ctx_id 0x55dc5151a4e0, pd_num 22, queues 5, queue_size 1024, vhca_id 0\n",
    "caps": "...",
    "send_engine": [
        "Send_engine 0: ...",
        "Send_engine 1: ...",
        ...
    ],
    "tables": [
        "Table 0x7fccd5dbfc40: ...",
        "Table 0x7fb453ad2520: ...",
        ...
    ]
}
```

Signed-off-by: Alexander Efimov <aefimov@nvidia.com>