HWS: Simplify dr_parse_table.fix_data
Remove the redundant checks for None values. `dict.get()` already has a
parameter for the default value if the key is not found in the
dictionary.

Note that this isn't exactly equivalent to the original if the
retrieved value really is `None`. However, this should not happen with
the current code.

Signed-off-by: Alexander Efimov <aefimov@nvidia.com>