HWS: Parallelize STE parsing
We now parse STEs on multiple cores, in 64k chunks. To keep output
deterministic, we buffer the resulting output before reassembling it in
order, but if the user values speed over determinism we expose the
--no_ordered_output flag.

This patch, together with writing output directly to the file, improves
STE parse speed by 7.5x. The numbers below do not reflect that because
the tool runtime also includes loading data from the disk, which is
something we have not optimized.

Measurements taken on a 4M STE dump, excluding FW dump time (so, with
--skip_dump):

			Wall time		Max RSS
Before			3:24			12.0G
This patch		1:00			 8.3G
This patch, no_ordered	0:57			 8.5G

Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>
Signed-off-by: Vlad Dogaru <vdogaru@nvidia.com>