HWS: Print output directly to file
Print to file instead of buffering a potentially huge string. In order
to keep output deterministic, we still need to buffer results from
individual chunks, but we try to eagerly print chunks which arrived in
order.

This brings down the parsing of a 4M STE dump from 1m2s to 55s.

Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>
Signed-off-by: Vlad Dogaru <vdogaru@nvidia.com>