HWS: Optimize FW STE resourcedump output parsing
Optimize FW STE resourcedump output parsing by introducing multi processing.
In a case of 4M table with 1M active STEs we saw parsing the output improved
from 12.7sec to 10.4sec which is ~16% improvement.
In a case of 4M table with 2M active STEs we saw parsing the output improved
from 18.1sec to 15.8sec which is ~13% improvement.

Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>