HWS: Parallelize resourcedump call and parsing
Call resource dump for each FW STE in parallel to parse the output of the
previous call.

In a case of dumping 2 STE tables from HW and parsing the output in parallel
we saw an improvement that one of the tables dumping from the HW was in parallel
to parsing the other table HW dump output parsing, and in this specific case
dumping and parsing the tables from HW improved from 100sec to 84sec.

Before:
	[Dump table A] => [Parse table A] => [Dump table B] => [Parse table B]
After:
	[Dump table A] => [Parse table A]
			  [Dump table B] => [Parse table B]

In the above case we saw a 16% improvement, since the case have 2 tables
which one of them is 4M entries and the other 512K entries, we did not
notice significant improvement, but in cases were there are more tables
we will notice more significant time improvement.

Signed-off-by: Hamdan Igbaria <hamdani@nvidia.com>